requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
user-agents==2.2.0
geoip2==4.7.0
validators==0.22.0
